Move code to /src
