+++ src/smiley_extractor.py
